+++ db/migrate/20200603134648_add_receipt_url_to_pay_charges.rb
@@ -1,5 +1,0 @@
-class AddReceiptUrlToPayCharges < ActiveRecord::Migration[4.2]
-  def change
-    add_column :pay_charges, :receipt_url, :string
-  end
-end