+++ lib/pay/paddle/subscription.rb
@@ -1,6 +1,15 @@
 module Pay
   module Paddle
     module Subscription
+      extend ActiveSupport::Concern
+
+      included do
+        scope :paddle, -> { where(processor: :paddle) }
+
+        store_accessor :data, :paddle_update_url
+        store_accessor :data, :paddle_cancel_url
+      end
+
       def paddle?
         processor == "paddle"
       end