+++ test/dummy/config/application.rb
@@ -11,7 +11,5 @@ module Dummy
     # Settings in config/environments/* take precedence over those specified here.
     # Application configuration should go into files in config/initializers
     # -- all .rb files in that directory are automatically loaded.
-
-    config.load_defaults 5.2
   end
 end