+++ test/dummy/config/routes.rb
@@ -18,8 +18,23 @@ Rails.application.routes.draw do
     end
   end
 
+  namespace :paddle do
+    resource :payment_method, namespace: :paddle
+    resources :subscriptions do
+      member do
+        patch :cancel
+        patch :resume
+      end
+    end
+    resources :charges do
+      member do
+        patch :refund
+      end
+    end
+  end
+
   namespace :stripe do
-    resource :payment_method, namespace: :braintree
+    resource :payment_method, namespace: :stripe
     resources :subscriptions do
       member do
         patch :cancel