+++ test/pay/paddle/webhooks/subscription_payment_succeeded_test.rb
@@ -15,7 +15,7 @@ class Pay::Paddle::Webhooks::SubscriptionPaymentSucceededTest < ActiveSupport::T
     charge = Pay.charge_model.last
     assert_equal Integer(@data["sale_gross"].to_f * 100), charge.amount
     assert_equal @data["payment_method"], charge.card_type
-    assert_equal @data["receipt_url"], charge.receipt_url
+    assert_equal @data["receipt_url"], charge.paddle_receipt_url
   end
 
   test "a charge isn't created if no corresponding user can be found" do